Rocha (aka ROKKA, aka Random Mocha)
Runs BDD unit tests using Mocha, but first randomizes the order of unit tests in each suite.

If the tests fail, the order is saved in the file to be used again on the next run, allowing you to debug the problem.
If the tests pass, the file with the saved order is deleted, and the tests will be reshuffled again on next run.

This demo shows running Rocha several times until one configuration shows failing assertion due to state pollution among the unit tests. Then we fix the suite using beforeEach block and rerun the tests again to confirm.

* Homepage: https://github.com/bahmutov/rocha ( https://github.com/bahmutov/rocha )
* NPM install npm install -g rocha
* Use rocha <spec pattern> , like rocha src/*-spec.js